directory: Add upload form.
Uploading works by POSTing a file to the directory. The original name
of the file is transmitted with the request and used as the file name.

We may later add a PUT request that is made against the full
destination URL and transmits only the file contents in the request
body.